Address PR changes
