+++ src/OpenTelemetry.Api/.publicApi/net452/PublicAPI.Unshipped.txt
@@ -170,7 +170,6 @@ OpenTelemetry.Trace.Tracer.StartSpan(string name, OpenTelemetry.Trace.SpanKind k
 OpenTelemetry.Trace.Tracer.StartSpan(string name, OpenTelemetry.Trace.SpanKind kind, in OpenTelemetry.Trace.TelemetrySpan parentSpan, OpenTelemetry.Trace.SpanAttributes initialAttributes = null, System.Collections.Generic.IEnumerable<OpenTelemetry.Trace.Link> links = null, System.DateTimeOffset startTime = default(System.DateTimeOffset)) -> OpenTelemetry.Trace.TelemetrySpan
 OpenTelemetry.Trace.TracerProvider
 OpenTelemetry.Trace.TracerProvider.GetTracer(string name, string version = null) -> OpenTelemetry.Trace.Tracer
-OpenTelemetry.Trace.TracerProvider.Shutdown(int timeoutMilliseconds = -1) -> bool
 OpenTelemetry.Trace.TracerProvider.TracerProvider() -> void
 abstract OpenTelemetry.Context.RuntimeContextSlot<T>.Get() -> T
 abstract OpenTelemetry.Context.RuntimeContextSlot<T>.Set(T value) -> void
@@ -309,4 +308,3 @@ virtual OpenTelemetry.Metrics.LabelSet.Labels.get -> System.Collections.Generic.
 virtual OpenTelemetry.Metrics.LabelSet.Labels.set -> void
 virtual OpenTelemetry.Metrics.MeterFactoryBase.GetMeter(string name, string version = null) -> OpenTelemetry.Metrics.Meter
 virtual OpenTelemetry.Metrics.MeterProvider.GetMeter(string name, string version = null) -> OpenTelemetry.Metrics.Meter
-virtual OpenTelemetry.Trace.TracerProvider.OnShutdown(int timeoutMilliseconds) -> bool
