+++ src/OpenTelemetry.Api/Trace/TracerProvider.cs
@@ -14,10 +14,7 @@
 // limitations under the License.
 // </copyright>
 
-using System;
 using System.Diagnostics;
-using System.Threading;
-using OpenTelemetry.Internal;
 
 namespace OpenTelemetry.Trace
 {
@@ -26,8 +23,6 @@ namespace OpenTelemetry.Trace
     /// </summary>
     public class TracerProvider : BaseProvider
     {
-        private int shutdownCount;
-
         /// <summary>
         /// Initializes a new instance of the <see cref="TracerProvider"/> class.
         /// </summary>
@@ -54,68 +49,6 @@ namespace OpenTelemetry.Trace
             }
 
             return new Tracer(new ActivitySource(name, version));
-        }
-
-        /// <summary>
-        /// Attempts to shutdown the processor, blocks the current thread until
-        /// shutdown completed or timed out.
-        /// </summary>
-        /// <param name="timeoutMilliseconds">
-        /// The number of milliseconds to wait, or <c>Timeout.Infinite</c> to
-        /// wait indefinitely.
-        /// </param>
-        /// <returns>
-        /// Returns <c>true</c> when shutdown succeeded; otherwise, <c>false</c>.
-        /// </returns>
-        /// <exception cref="System.ArgumentOutOfRangeException">
-        /// Thrown when the <c>timeoutMilliseconds</c> is smaller than -1.
-        /// </exception>
-        /// <remarks>
-        /// This function guarantees thread-safety. Only the first call will
-        /// win, subsequent calls will be no-op.
-        /// </remarks>
-        public bool Shutdown(int timeoutMilliseconds = Timeout.Infinite)
-        {
-            if (timeoutMilliseconds < 0 && timeoutMilliseconds != Timeout.Infinite)
-            {
-                throw new ArgumentOutOfRangeException(nameof(timeoutMilliseconds), timeoutMilliseconds, "timeoutMilliseconds should be non-negative.");
-            }
-
-            if (Interlocked.Increment(ref this.shutdownCount) > 1)
-            {
-                return false; // shutdown already called
-            }
-
-            try
-            {
-                return this.OnShutdown(timeoutMilliseconds);
-            }
-            catch (Exception ex)
-            {
-                OpenTelemetryApiEventSource.Log.TracerProviderException(nameof(this.Shutdown), ex);
-                return false;
-            }
-        }
-
-        /// <summary>
-        /// Called by <c>Shutdown</c>. This function should block the current
-        /// thread until shutdown completed or timed out.
-        /// </summary>
-        /// <param name="timeoutMilliseconds">
-        /// The number of milliseconds to wait, or <c>Timeout.Infinite</c> to
-        /// wait indefinitely.
-        /// </param>
-        /// <returns>
-        /// Returns <c>true</c> when shutdown succeeded; otherwise, <c>false</c>.
-        /// </returns>
-        /// <remarks>
-        /// This function is called synchronously on the thread which made the
-        /// first call to <c>Shutdown</c>. This function should not throw
-        /// exceptions.
-        /// </remarks>
-        protected virtual bool OnShutdown(int timeoutMilliseconds)
-        {
-            return true;
         }
     }
 }
