+++ src/OpenTelemetry/Trace/TracerProviderExtensions.cs
@@ -16,6 +16,8 @@
 
 using System;
 using System.Diagnostics;
+using System.Threading;
+using OpenTelemetry.Internal;
 
 namespace OpenTelemetry.Trace
 {
@@ -39,6 +41,58 @@ namespace OpenTelemetry.Trace
             }
 
             return provider;
+        }
+
+        /// <summary>
+        /// Attempts to shutdown the TracerProviderSdk, blocks the current thread until
+        /// shutdown completed or timed out.
+        /// </summary>
+        /// <param name="provider">TracerProviderSdk instance on which Shutdown will be called.</param>
+        /// <param name="timeoutMilliseconds">
+        /// The number of milliseconds to wait, or <c>Timeout.Infinite</c> to
+        /// wait indefinitely.
+        /// </param>
+        /// <returns>
+        /// Returns <c>true</c> when shutdown succeeded; otherwise, <c>false</c>.
+        /// </returns>
+        /// <exception cref="System.ArgumentOutOfRangeException">
+        /// Thrown when the <c>timeoutMilliseconds</c> is smaller than -1.
+        /// </exception>
+        /// <remarks>
+        /// This function guarantees thread-safety. Only the first call will
+        /// win, subsequent calls will be no-op.
+        /// </remarks>
+        public static bool Shutdown(this TracerProvider provider, int timeoutMilliseconds = Timeout.Infinite)
+        {
+            if (provider == null)
+            {
+                throw new ArgumentNullException(nameof(provider));
+            }
+
+            if (provider is TracerProviderSdk tracerProviderSdk)
+            {
+                if (timeoutMilliseconds < 0 && timeoutMilliseconds != Timeout.Infinite)
+                {
+                    throw new ArgumentOutOfRangeException(nameof(timeoutMilliseconds), timeoutMilliseconds, "timeoutMilliseconds should be non-negative.");
+                }
+
+                if (Interlocked.Increment(ref tracerProviderSdk.ShutdownCount) > 1)
+                {
+                    return false; // shutdown already called
+                }
+
+                try
+                {
+                    return tracerProviderSdk.OnShutdown(timeoutMilliseconds);
+                }
+                catch (Exception ex)
+                {
+                    OpenTelemetrySdkEventSource.Log.TracerProviderException(nameof(tracerProviderSdk.OnShutdown), ex);
+                    return false;
+                }
+            }
+
+            return true;
         }
     }
 }
